+++ app.py
@@ -1,6 +1,6 @@
 import os
 
-from hflossk.run import app
+from hflossk.site import app
 
 if __name__ == "__main__":
     if 'OPENSHIFT_PYTHON_IP' in os.environ:
